Add aggregation_terms colours to each log line as well
This may help distinguishing different kinds of log lines quickly, e.g.
`level=ERROR` or `status.code=500`, which will have a red bar on the
left of the log line.

It also makes the logs a bit more fun!  :)
